simplify reading
New to FP, I was originally confused by the fact that we had two things named `Person`in lesson 2, then suddenly switched to having two different names in lesson 3. I also was confused about what these things were. It tooks reading through questions on Discord to clarify that one is a Type, the other a constructor, and therefor a `(String,Int,Bool) -> Person` function.

This tries to explicit things a little (although I agree a comment might
not be the best way to do this).